refactor: migrate dashboard  to modular arc
- Migrate Dashboard module
